working on new rna abstractor
