+++ graphlearn/abstract_graphs/directedgraphtools.py
@@ -1,0 +1,35 @@
+
+
+
+def get_start_and_end_node(graph):
+    # make n the first node of the sequence
+    start=-1
+    end=-1
+    for n,d in graph.nodes_iter(data=True):
+
+        # edge nodes cant be start or end
+        if 'edge' in d:
+            continue
+
+        # check for start
+        if start == -1:
+            l= graph.predecessors(n)
+            if len(l)==0:
+                start = n
+            if len(l)==1:
+                if graph.node[ l[0] ]['label']=='=':
+                    start = n
+
+        # check for end:
+        if end == -1:
+            l= graph.neighbors(n)
+            if len(l)==0:
+                end = n
+            if len(l)==1:
+                if graph.node[ l[0] ]['label']=='=':
+                    end = n
+
+    # check and return
+    if start==-1 or end==-1:
+        raise Exception ('your beautiful "rna" has no clear start or end')
+    return start,end
